+++ packages/node/test/onunhandledrejection.test.ts
@@ -1,15 +1,29 @@
-import { Hub, init, Integrations, Scope } from '../src';
+import { Scope } from '@sentry/core';
+import { Hub } from '@sentry/hub';
 
-const dsn = 'https://53039209a22b4ec1bcc296a3c9fdecd6@sentry.io/4291';
+import { OnUnhandledRejection } from '../src/integrations/onunhandledrejection';
+
+jest.mock('@sentry/hub', () => {
+  // we just want to short-circuit it, so dont worry about types
+  const original = jest.requireActual('@sentry/hub');
+  original.Hub.prototype.getIntegration = () => true;
+  return {
+    ...original,
+    getCurrentHub: () => new Hub(),
+  };
+});
 
 describe('unhandled promises', () => {
   test('install global listener', () => {
-    init({ dsn });
+    const integration = new OnUnhandledRejection();
+    integration.setupOnce();
     expect(process.listeners('unhandledRejection')).toHaveLength(1);
   });
 
   test('sendUnhandledPromise', () => {
-    const integration = new Integrations.OnUnhandledRejection();
+    const integration = new OnUnhandledRejection();
+    integration.setupOnce();
+
     const promise = {
       domain: {
         sentryContext: {
@@ -28,6 +42,9 @@ describe('unhandled promises', () => {
 
     integration.sendUnhandledPromise('bla', promise);
 
+    expect(captureException.mock.calls[0][1]?.data).toEqual({
+      mechanism: { handled: false, type: 'onunhandledrejection' },
+    });
     expect(captureException.mock.calls[0][0]).toBe('bla');
     expect(setUser.mock.calls[0][0]).toEqual({ id: 1 });
     expect(setExtra.mock.calls[0]).toEqual(['unhandledPromiseRejection', true]);