+++ packages/node/test/transports/http.test.ts
@@ -1,4 +1,5 @@
-import { TransportOptions } from '@sentry/types';
+import { Session } from '@sentry/hub';
+import { Event, SessionStatus, Status, TransportOptions } from '@sentry/types';
 import { SentryError } from '@sentry/utils';
 import * as http from 'http';
 import * as HttpsProxyAgent from 'https-proxy-agent';
@@ -7,7 +8,29 @@ import { HTTPTransport } from '../../src/transports/http';
 
 const mockSetEncoding = jest.fn();
 const dsn = 'http://9e9fd4523d784609a5fc0ebb1080592f@sentry.io:8989/mysubpath/50622';
-const transportPath = '/mysubpath/api/50622/store/';
+const storePath = '/mysubpath/api/50622/store/';
+const envelopePath = '/mysubpath/api/50622/envelope/';
+const eventPayload: Event = {
+  event_id: '1337',
+};
+const transactionPayload: Event = {
+  event_id: '42',
+  type: 'transaction',
+};
+const sessionPayload: Session = {
+  environment: 'test',
+  release: '1.0',
+  sid: '353463243253453254',
+  errors: 0,
+  started: Date.now(),
+  timestamp: Date.now(),
+  init: true,
+  duration: 0,
+  status: SessionStatus.Exited,
+  update: jest.fn(),
+  close: jest.fn(),
+  toJSON: jest.fn(),
+};
 let mockReturnCode = 200;
 let mockHeaders = {};
 
@@ -28,12 +51,12 @@ function createTransport(options: TransportOptions): HTTPTransport {
   return transport;
 }
 
-function assertBasicOptions(options: any): void {
+function assertBasicOptions(options: any, useEnvelope: boolean = false): void {
   expect(options.headers['X-Sentry-Auth']).toContain('sentry_version');
   expect(options.headers['X-Sentry-Auth']).toContain('sentry_client');
   expect(options.headers['X-Sentry-Auth']).toContain('sentry_key');
   expect(options.port).toEqual('8989');
-  expect(options.path).toEqual(transportPath);
+  expect(options.path).toEqual(useEnvelope ? envelopePath : storePath);
   expect(options.hostname).toEqual('sentry.io');
 }
 
@@ -66,6 +89,28 @@ describe('HTTPTransport', () => {
     } catch (e) {
       const requestOptions = (transport.module!.request as jest.Mock).mock.calls[0][0];
       assertBasicOptions(requestOptions);
+      expect(e).toEqual(new SentryError(`HTTP Error (${mockReturnCode})`));
+    }
+  });
+
+  test('send 200 session', async () => {
+    const transport = createTransport({ dsn });
+    await transport.sendSession(new Session());
+
+    const requestOptions = (transport.module!.request as jest.Mock).mock.calls[0][0];
+    assertBasicOptions(requestOptions, true);
+    expect(mockSetEncoding).toHaveBeenCalled();
+  });
+
+  test('send 400 session', async () => {
+    mockReturnCode = 400;
+    const transport = createTransport({ dsn });
+
+    try {
+      await transport.sendSession(new Session());
+    } catch (e) {
+      const requestOptions = (transport.module!.request as jest.Mock).mock.calls[0][0];
+      assertBasicOptions(requestOptions, true);
       expect(e).toEqual(new SentryError(`HTTP Error (${mockReturnCode})`));
     }
   });
@@ -117,9 +162,12 @@ describe('HTTPTransport', () => {
     try {
       await transport.sendEvent({ message: 'test' });
     } catch (e) {
-      expect(e).toEqual(
-        new SentryError(`Transport locked till ${new Date(now + retryAfterSeconds * 1000)} due to too many requests.`),
+      expect(e.status).toEqual(429);
+      expect(e.reason).toEqual(
+        `Transport locked till ${new Date(now + retryAfterSeconds * 1000)} due to too many requests.`,
       );
+      expect(e.payload.message).toEqual('test');
+      expect(e.type).toEqual('event');
     }
 
     try {
@@ -127,6 +175,388 @@ describe('HTTPTransport', () => {
     } catch (e) {
       expect(e).toEqual(new SentryError(`HTTP Error (${mockReturnCode})`));
     }
+
+    mock.mockRestore();
+  });
+
+  test('back-off using x-sentry-rate-limits with bogus headers and missing categories should just lock them all', async () => {
+    const retryAfterSeconds = 60;
+    mockReturnCode = 429;
+    mockHeaders = {
+      'x-sentry-rate-limits': `sgthrthewhertht`,
+    };
+    const transport = createTransport({ dsn });
+    const now = Date.now();
+    const mock = jest
+      .spyOn(Date, 'now')
+      // 1st event - _isRateLimited - false
+      .mockReturnValueOnce(now)
+      // 1st event - _handleRateLimit
+      .mockReturnValueOnce(now)
+      // 2nd event - _isRateLimited - true (event category)
+      .mockReturnValueOnce(now + (retryAfterSeconds / 2) * 1000)
+      // 3rd event - _isRateLimited - true (transaction category)
+      .mockReturnValueOnce(now + (retryAfterSeconds / 2) * 1000)
+      // 4th event - _isRateLimited - false (event category)
+      .mockReturnValueOnce(now + retryAfterSeconds * 1000)
+      // 4th event - _handleRateLimit
+      .mockReturnValueOnce(now + retryAfterSeconds * 1000)
+      // 5th event - _isRateLimited - false (transaction category)
+      .mockReturnValueOnce(now + retryAfterSeconds * 1000)
+      // 5th event - _handleRateLimit
+      .mockReturnValueOnce(now + retryAfterSeconds * 1000);
+
+    try {
+      await transport.sendEvent(eventPayload);
+    } catch (e) {
+      expect(e).toEqual(new SentryError(`HTTP Error (${mockReturnCode})`));
+    }
+
+    try {
+      await transport.sendEvent(eventPayload);
+    } catch (e) {
+      expect(e.status).toEqual(429);
+      expect(e.reason).toEqual(
+        `Transport locked till ${new Date(now + retryAfterSeconds * 1000)} due to too many requests.`,
+      );
+      expect(e.payload).toEqual(eventPayload);
+      expect(e.type).toEqual('event');
+    }
+
+    try {
+      await transport.sendEvent(transactionPayload);
+    } catch (e) {
+      expect(e.status).toEqual(429);
+      expect(e.reason).toEqual(
+        `Transport locked till ${new Date(now + retryAfterSeconds * 1000)} due to too many requests.`,
+      );
+      expect(e.payload).toEqual(transactionPayload);
+      expect(e.type).toEqual('transaction');
+    }
+
+    mockHeaders = {};
+    mockReturnCode = 200;
+
+    const eventRes = await transport.sendEvent(eventPayload);
+    expect(eventRes.status).toEqual(Status.Success);
+
+    const transactionRes = await transport.sendEvent(transactionPayload);
+    expect(transactionRes.status).toEqual(Status.Success);
+
+    mock.mockRestore();
+  });
+
+  test('back-off using x-sentry-rate-limits with single category', async () => {
+    const retryAfterSeconds = 10;
+    mockReturnCode = 429;
+    mockHeaders = {
+      'x-sentry-rate-limits': `${retryAfterSeconds}:error:scope`,
+    };
+    const transport = createTransport({ dsn });
+    const now = Date.now();
+    const mock = jest
+      .spyOn(Date, 'now')
+      // 1st event - _isRateLimited - false
+      .mockReturnValueOnce(now)
+      // 1st event - _handleRateLimit
+      .mockReturnValueOnce(now)
+      // 2nd event - _isRateLimited - false (different category)
+      .mockReturnValueOnce(now + (retryAfterSeconds / 2) * 1000)
+      // 2nd event - _handleRateLimit
+      .mockReturnValueOnce(now + (retryAfterSeconds / 2) * 1000)
+      // 3rd event - _isRateLimited - false (different category - sessions)
+      .mockReturnValueOnce(now + (retryAfterSeconds / 2) * 1000)
+      // 3rd event - _handleRateLimit
+      .mockReturnValueOnce(now + (retryAfterSeconds / 2) * 1000)
+      // 4th event - _isRateLimited - true
+      .mockReturnValueOnce(now + (retryAfterSeconds / 2) * 1000)
+      // 5th event - _isRateLimited - false
+      .mockReturnValueOnce(now + retryAfterSeconds * 1000)
+      // 5th event - _handleRateLimit
+      .mockReturnValueOnce(now + retryAfterSeconds * 1000);
+
+    try {
+      await transport.sendEvent(eventPayload);
+    } catch (e) {
+      expect(e).toEqual(new SentryError(`HTTP Error (${mockReturnCode})`));
+    }
+
+    mockHeaders = {};
+    mockReturnCode = 200;
+
+    const transactionRes = await transport.sendEvent(transactionPayload);
+    expect(transactionRes.status).toEqual(Status.Success);
+
+    const sessionsRes = await transport.sendSession(sessionPayload);
+    expect(sessionsRes.status).toEqual(Status.Success);
+
+    try {
+      await transport.sendEvent(eventPayload);
+    } catch (e) {
+      expect(e.status).toEqual(429);
+      expect(e.reason).toEqual(
+        `Transport locked till ${new Date(now + retryAfterSeconds * 1000)} due to too many requests.`,
+      );
+      expect(e.payload).toEqual(eventPayload);
+      expect(e.type).toEqual('event');
+    }
+
+    const eventRes = await transport.sendEvent(eventPayload);
+    expect(eventRes.status).toEqual(Status.Success);
+
+    mock.mockRestore();
+  });
+
+  test('back-off using x-sentry-rate-limits with multiple category', async () => {
+    const retryAfterSeconds = 10;
+    mockReturnCode = 429;
+    mockHeaders = {
+      'x-sentry-rate-limits': `${retryAfterSeconds}:error;transaction;session:scope`,
+    };
+    const transport = createTransport({ dsn });
+    const now = Date.now();
+    const mock = jest
+      .spyOn(Date, 'now')
+      // 1st event - _isRateLimited - false
+      .mockReturnValueOnce(now)
+      // 1st event - _handleRateLimit
+      .mockReturnValueOnce(now)
+      // 2nd event - _isRateLimited - true (event category)
+      .mockReturnValueOnce(now + (retryAfterSeconds / 2) * 1000)
+      // 3rd event - _isRateLimited - true (sessions category)
+      .mockReturnValueOnce(now + (retryAfterSeconds / 2) * 1000)
+      // 4th event - _isRateLimited - true (transactions category)
+      .mockReturnValueOnce(now + (retryAfterSeconds / 2) * 1000)
+      // 5th event - _isRateLimited - false (event category)
+      .mockReturnValueOnce(now + retryAfterSeconds * 1000)
+      // 5th event - _handleRateLimit
+      .mockReturnValueOnce(now + retryAfterSeconds * 1000)
+      // 6th event - _isRateLimited - false (sessions category)
+      .mockReturnValueOnce(now + retryAfterSeconds * 1000)
+      // 6th event - handleRateLimit
+      .mockReturnValueOnce(now + retryAfterSeconds * 1000)
+      // 7th event - _isRateLimited - false (transaction category)
+      .mockReturnValueOnce(now + retryAfterSeconds * 1000)
+      // 7th event - handleRateLimit
+      .mockReturnValueOnce(now + retryAfterSeconds * 1000);
+
+    try {
+      await transport.sendEvent(eventPayload);
+    } catch (e) {
+      expect(e).toEqual(new SentryError(`HTTP Error (${mockReturnCode})`));
+    }
+
+    try {
+      await transport.sendEvent(eventPayload);
+    } catch (e) {
+      expect(e.status).toEqual(429);
+      expect(e.reason).toEqual(
+        `Transport locked till ${new Date(now + retryAfterSeconds * 1000)} due to too many requests.`,
+      );
+      expect(e.payload).toEqual(eventPayload);
+      expect(e.type).toEqual('event');
+    }
+
+    try {
+      await transport.sendSession(sessionPayload);
+    } catch (e) {
+      expect(e.status).toEqual(429);
+      expect(e.reason).toEqual(
+        `Transport locked till ${new Date(now + retryAfterSeconds * 1000)} due to too many requests.`,
+      );
+      expect(e.payload.environment).toEqual(sessionPayload.environment);
+      expect(e.payload.release).toEqual(sessionPayload.release);
+      expect(e.payload.sid).toEqual(sessionPayload.sid);
+      expect(e.type).toEqual('session');
+    }
+
+    try {
+      await transport.sendEvent(transactionPayload);
+    } catch (e) {
+      expect(e.status).toEqual(429);
+      expect(e.reason).toEqual(
+        `Transport locked till ${new Date(now + retryAfterSeconds * 1000)} due to too many requests.`,
+      );
+      expect(e.payload).toEqual(transactionPayload);
+      expect(e.type).toEqual('transaction');
+    }
+
+    mockHeaders = {};
+    mockReturnCode = 200;
+
+    const eventRes = await transport.sendEvent(eventPayload);
+    expect(eventRes.status).toEqual(Status.Success);
+
+    const sessionsRes = await transport.sendSession(sessionPayload);
+    expect(sessionsRes.status).toEqual(Status.Success);
+
+    const transactionRes = await transport.sendEvent(transactionPayload);
+    expect(transactionRes.status).toEqual(Status.Success);
+
+    mock.mockRestore();
+  });
+
+  test('back-off using x-sentry-rate-limits with missing categories should lock them all', async () => {
+    const retryAfterSeconds = 10;
+    mockReturnCode = 429;
+    mockHeaders = {
+      'x-sentry-rate-limits': `${retryAfterSeconds}::scope`,
+    };
+    const transport = createTransport({ dsn });
+    const now = Date.now();
+    const mock = jest
+      .spyOn(Date, 'now')
+      // 1st event - _isRateLimited - false
+      .mockReturnValueOnce(now)
+      // 1st event - _handleRateLimit
+      .mockReturnValueOnce(now)
+      // 2nd event - _isRateLimited - true (event category)
+      .mockReturnValueOnce(now + (retryAfterSeconds / 2) * 1000)
+      // 3rd event - _isRateLimited - true (transaction category)
+      .mockReturnValueOnce(now + (retryAfterSeconds / 2) * 1000)
+      // 4th event - _isRateLimited - false (event category)
+      .mockReturnValueOnce(now + retryAfterSeconds * 1000)
+      // 4th event - _handleRateLimit
+      .mockReturnValueOnce(now + retryAfterSeconds * 1000)
+      // 5th event - _isRateLimited - false (transaction category)
+      .mockReturnValueOnce(now + retryAfterSeconds * 1000)
+      // 5th event - _handleRateLimit
+      .mockReturnValueOnce(now + retryAfterSeconds * 1000);
+
+    try {
+      await transport.sendEvent(eventPayload);
+    } catch (e) {
+      expect(e).toEqual(new SentryError(`HTTP Error (${mockReturnCode})`));
+    }
+
+    try {
+      await transport.sendEvent(eventPayload);
+    } catch (e) {
+      expect(e.status).toEqual(429);
+      expect(e.reason).toEqual(
+        `Transport locked till ${new Date(now + retryAfterSeconds * 1000)} due to too many requests.`,
+      );
+      expect(e.payload).toEqual(eventPayload);
+      expect(e.type).toEqual('event');
+    }
+
+    try {
+      await transport.sendEvent(transactionPayload);
+    } catch (e) {
+      expect(e.status).toEqual(429);
+      expect(e.reason).toEqual(
+        `Transport locked till ${new Date(now + retryAfterSeconds * 1000)} due to too many requests.`,
+      );
+      expect(e.payload).toEqual(transactionPayload);
+      expect(e.type).toEqual('transaction');
+    }
+
+    mockHeaders = {};
+    mockReturnCode = 200;
+
+    const eventRes = await transport.sendEvent(eventPayload);
+    expect(eventRes.status).toEqual(Status.Success);
+
+    const transactionRes = await transport.sendEvent(transactionPayload);
+    expect(transactionRes.status).toEqual(Status.Success);
+
+    mock.mockRestore();
+  });
+
+  test('back-off using x-sentry-rate-limits with bogus categories should be dropped', async () => {
+    const retryAfterSeconds = 10;
+    mockReturnCode = 429;
+    mockHeaders = {
+      'x-sentry-rate-limits': `${retryAfterSeconds}:error;safegreg;eqwerw:scope`,
+    };
+    const transport = createTransport({ dsn });
+    const now = Date.now();
+    const mock = jest
+      .spyOn(Date, 'now')
+      // 1st event - _isRateLimited - false
+      .mockReturnValueOnce(now)
+      // 1st event - _handleRateLimit
+      .mockReturnValueOnce(now)
+      // 2nd event - _isRateLimited - true (event category)
+      .mockReturnValueOnce(now + (retryAfterSeconds / 2) * 1000)
+      // 3rd event - _isRateLimited - false (transaction category)
+      .mockReturnValueOnce(now + (retryAfterSeconds / 2) * 1000)
+      // 3rd Event - _handleRateLimit
+      .mockReturnValueOnce(now + (retryAfterSeconds / 2) * 1000)
+      // 4th event - _isRateLimited - false (event category)
+      .mockReturnValueOnce(now + retryAfterSeconds * 1000)
+      // 4th event - _handleRateLimit
+      .mockReturnValueOnce(now + retryAfterSeconds * 1000);
+
+    try {
+      await transport.sendEvent(eventPayload);
+    } catch (e) {
+      expect(e).toEqual(new SentryError(`HTTP Error (${mockReturnCode})`));
+    }
+
+    try {
+      await transport.sendEvent(eventPayload);
+    } catch (e) {
+      expect(e.status).toEqual(429);
+      expect(e.reason).toEqual(
+        `Transport locked till ${new Date(now + retryAfterSeconds * 1000)} due to too many requests.`,
+      );
+      expect(e.payload).toEqual(eventPayload);
+      expect(e.type).toEqual('event');
+    }
+
+    mockHeaders = {};
+    mockReturnCode = 200;
+
+    const transactionRes = await transport.sendEvent(transactionPayload);
+    expect(transactionRes.status).toEqual(Status.Success);
+
+    const eventRes = await transport.sendEvent(eventPayload);
+    expect(eventRes.status).toEqual(Status.Success);
+
+    mock.mockRestore();
+  });
+
+  test('back-off using x-sentry-rate-limits should also trigger for 200 responses', async () => {
+    const retryAfterSeconds = 10;
+    mockReturnCode = 200;
+    mockHeaders = {
+      'x-sentry-rate-limits': `${retryAfterSeconds}:error;transaction:scope`,
+    };
+    const transport = createTransport({ dsn });
+    const now = Date.now();
+    const mock = jest
+      .spyOn(Date, 'now')
+      // 1st event - _isRateLimited - false
+      .mockReturnValueOnce(now)
+      // 1st event - _handleRateLimit
+      .mockReturnValueOnce(now)
+      // 2nd event - _isRateLimited - true
+      .mockReturnValueOnce(now + (retryAfterSeconds / 2) * 1000)
+      // 3rd event - _isRateLimited - false
+      .mockReturnValueOnce(now + retryAfterSeconds * 1000)
+      // 3rd event - _handleRateLimit
+      .mockReturnValueOnce(now + retryAfterSeconds * 1000);
+
+    let eventRes = await transport.sendEvent(eventPayload);
+    expect(eventRes.status).toEqual(Status.Success);
+
+    try {
+      await transport.sendEvent(eventPayload);
+    } catch (e) {
+      expect(e.status).toEqual(429);
+      expect(e.reason).toEqual(
+        `Transport locked till ${new Date(now + retryAfterSeconds * 1000)} due to too many requests.`,
+      );
+      expect(e.payload).toEqual(eventPayload);
+      expect(e.type).toEqual('event');
+    }
+
+    mockReturnCode = 200;
+    mockHeaders = {};
+
+    eventRes = await transport.sendEvent(eventPayload);
+    expect(eventRes.status).toEqual(Status.Success);
 
     mock.mockRestore();
   });